Correção na paginação, e adicionado confirmação de exclusão

+++ meus-anuncios.php
@@ -48,7 +48,7 @@ if (empty($_SESSION['clogin'])) {
 			<td><?php echo $anuncio['titulo'] ?></td>
 			<td>R$ <?php echo number_format($anuncio['valor'], 2) ?></td>
 			<td><a href="editar-anuncio.php?id=<?php echo $anuncio['id_anuncio'] ?>"><img src="assets/img/editar.png" height="23" data-toggle="tooltip" title="editar anúncio"></a></td>
-			<td><a href="excluir-anuncio.php?id=<?php echo $anuncio['id_anuncio'] ?>"><img src="assets/img/excluir.png" height="23" data-toggle="tooltip" title="excluir anúncio"></a></td>
+			<td><a href="excluir-anuncio.php?id=<?php echo $anuncio['id_anuncio'] ?>"><img src="assets/img/excluir.png" height="23" data-toggle="tooltip" title="excluir anúncio" onClick="return confirm('Deseja realmente excluir registro?')"></a></td>
 		</tr>
 		<?php
 		} // finaliza foreach
